core: move syscall name mapping out of context
