+++ crates/sandlock-core/src/context.rs
@@ -119,127 +119,7 @@ pub(crate) fn read_u32_fd(fd: RawFd) -> io::Result<u32> {
     Ok(u32::from_le_bytes(buf))
 }
 
-// ============================================================
-// Syscall name → number mapping
-// ============================================================
-
-/// Map a syscall name to its `libc::SYS_*` number.
-///
-/// Covers all names in `DEFAULT_BLOCKLIST_SYSCALLS` plus extras needed for
-/// notif and arg-filter lists.
-pub fn syscall_name_to_nr(name: &str) -> Option<u32> {
-    let nr: i64 = match name {
-        "mount" => libc::SYS_mount,
-        "umount2" => libc::SYS_umount2,
-        "pivot_root" => libc::SYS_pivot_root,
-        "swapon" => libc::SYS_swapon,
-        "swapoff" => libc::SYS_swapoff,
-        "reboot" => libc::SYS_reboot,
-        "sethostname" => libc::SYS_sethostname,
-        "setdomainname" => libc::SYS_setdomainname,
-        "kexec_load" => libc::SYS_kexec_load,
-        "init_module" => libc::SYS_init_module,
-        "finit_module" => libc::SYS_finit_module,
-        "delete_module" => libc::SYS_delete_module,
-        "unshare" => libc::SYS_unshare,
-        "setns" => libc::SYS_setns,
-        "perf_event_open" => libc::SYS_perf_event_open,
-        "bpf" => libc::SYS_bpf,
-        "userfaultfd" => libc::SYS_userfaultfd,
-        "keyctl" => libc::SYS_keyctl,
-        "add_key" => libc::SYS_add_key,
-        "request_key" => libc::SYS_request_key,
-        "ptrace" => libc::SYS_ptrace,
-        "process_vm_readv" => libc::SYS_process_vm_readv,
-        "process_vm_writev" => libc::SYS_process_vm_writev,
-        "open_by_handle_at" => libc::SYS_open_by_handle_at,
-        "name_to_handle_at" => libc::SYS_name_to_handle_at,
-        "ioperm" => arch::SYS_IOPERM?,
-        "iopl" => arch::SYS_IOPL?,
-        "quotactl" => libc::SYS_quotactl,
-        "acct" => libc::SYS_acct,
-        "lookup_dcookie" => libc::SYS_lookup_dcookie,
-        // nfsservctl was removed in Linux 3.1; no libc constant — skip
-        "personality" => libc::SYS_personality,
-        "io_uring_setup" => libc::SYS_io_uring_setup,
-        "io_uring_enter" => libc::SYS_io_uring_enter,
-        "io_uring_register" => libc::SYS_io_uring_register,
-        // Additional syscalls for notif/arg filters
-        "clone" => libc::SYS_clone,
-        "clone3" => libc::SYS_clone3,
-        "vfork" => arch::SYS_VFORK?,
-        "mmap" => libc::SYS_mmap,
-        "munmap" => libc::SYS_munmap,
-        "brk" => libc::SYS_brk,
-        "mremap" => libc::SYS_mremap,
-        "connect" => libc::SYS_connect,
-        "sendto" => libc::SYS_sendto,
-        "sendmsg" => libc::SYS_sendmsg,
-        "sendmmsg" => libc::SYS_sendmmsg,
-        "ioctl" => libc::SYS_ioctl,
-        "socket" => libc::SYS_socket,
-        "prctl" => libc::SYS_prctl,
-        "getrandom" => libc::SYS_getrandom,
-        "openat" => libc::SYS_openat,
-        "open" => arch::SYS_OPEN?,
-        "getdents64" => libc::SYS_getdents64,
-        "getdents" => arch::SYS_GETDENTS?,
-        "bind" => libc::SYS_bind,
-        "getsockname" => libc::SYS_getsockname,
-        "clock_gettime" => libc::SYS_clock_gettime,
-        "gettimeofday" => libc::SYS_gettimeofday,
-        "time" => arch::SYS_TIME?,
-        "clock_nanosleep" => libc::SYS_clock_nanosleep,
-        "timerfd_settime" => libc::SYS_timerfd_settime,
-        "timer_settime" => libc::SYS_timer_settime,
-        "execve" => libc::SYS_execve,
-        "execveat" => libc::SYS_execveat,
-        // COW filesystem syscalls
-        "unlinkat" => libc::SYS_unlinkat,
-        "mkdirat" => libc::SYS_mkdirat,
-        "renameat2" => libc::SYS_renameat2,
-        "newfstatat" => libc::SYS_newfstatat,
-        "statx" => libc::SYS_statx,
-        "faccessat" => libc::SYS_faccessat,
-        "symlinkat" => libc::SYS_symlinkat,
-        "linkat" => libc::SYS_linkat,
-        "fchmodat" => libc::SYS_fchmodat,
-        "fchownat" => libc::SYS_fchownat,
-        "readlinkat" => libc::SYS_readlinkat,
-        "truncate" => libc::SYS_truncate,
-        "utimensat" => libc::SYS_utimensat,
-        "unlink" => arch::SYS_UNLINK?,
-        "rmdir" => arch::SYS_RMDIR?,
-        "mkdir" => arch::SYS_MKDIR?,
-        "rename" => arch::SYS_RENAME?,
-        "stat" => arch::SYS_STAT?,
-        "lstat" => arch::SYS_LSTAT?,
-        "access" => arch::SYS_ACCESS?,
-        "symlink" => arch::SYS_SYMLINK?,
-        "link" => arch::SYS_LINK?,
-        "chmod" => arch::SYS_CHMOD?,
-        "chown" => arch::SYS_CHOWN?,
-        "lchown" => arch::SYS_LCHOWN?,
-        "readlink" => arch::SYS_READLINK?,
-        "futimesat" => arch::SYS_FUTIMESAT?,
-        "fork" => arch::SYS_FORK?,
-        // SysV IPC (gated by extra_allow_syscalls=["sysv_ipc"]; denied by default)
-        "shmget" => libc::SYS_shmget,
-        "shmat" => libc::SYS_shmat,
-        "shmdt" => libc::SYS_shmdt,
-        "shmctl" => libc::SYS_shmctl,
-        "msgget" => libc::SYS_msgget,
-        "msgsnd" => libc::SYS_msgsnd,
-        "msgrcv" => libc::SYS_msgrcv,
-        "msgctl" => libc::SYS_msgctl,
-        "semget" => libc::SYS_semget,
-        "semop" => libc::SYS_semop,
-        "semctl" => libc::SYS_semctl,
-        "semtimedop" => libc::SYS_semtimedop,
-        _ => return None,
-    };
-    Some(nr as u32)
-}
+use crate::seccomp::syscall_names::syscall_name_to_nr;
 
 // ============================================================
 // Sandbox → syscall lists
