+++ crates/sandlock-core/src/seccomp/mod.rs
@@ -4,3 +4,4 @@ pub mod dispatch;
 pub mod notif;
 pub(crate) mod state;
 pub mod syscall;
+pub mod syscall_names;
